-Prototype-UP Paper Saftey MGT System/requirements.txt
mysql-connector-python
mysqlclient
pyodbc
